TaskSwitcher: add option to show the process name before the task name
Add a new proc_name_first option which is used to configure whether the process name must be shown before or after the task name.

Enabling this option makes Keypirinha  sort the windows belonging to the same application together.

Another advantage of enabling this option is that it also makes it easier to search for the windows belonging to an app first, and then narrow the search down by typing part of the window title.

The option is disabled by default (i.e. the original behavior is maintained).